feat(PCI): Refactoring in progress
