+++ kernel/src/io/usb.rs
@@ -25,7 +25,9 @@ pub fn init() {
         .iter()
         .filter(|device| device.csp().prog_if() == 0x30)
         .filter_map(|device| {
-            if let Some(Bar::Memory(memory_bar)) = device.bar(0) {
+            if let Some(Bar::Memory(memory_bar)) =
+                crate::pci::with_pci_handler(|handler| handler.read_bar(device, 0))
+            {
                 Some(memory_bar.base_address())
             } else {
                 None
