+++ kernel/src/mem/page_alloc/pmap.rs
@@ -4,7 +4,9 @@
 
 use x86_64::{
     PhysAddr, VirtAddr,
-    structures::paging::{Mapper, Page, PageSize, PageTableFlags, PhysFrame, Size4KiB},
+    structures::paging::{
+        Mapper, Page, PageSize, PageTableFlags, PhysFrame, RecursivePageTable, Size4KiB,
+    },
 };
 
 use crate::mem::{frame_alloc, page_alloc, page_table};
@@ -14,29 +16,38 @@ use crate::mem::{frame_alloc, page_alloc, page_table};
 ///
 /// Be careful to only used the original mapped length, as accessing outside
 /// could result in undefined behavior if the memory is used by another mapping.
-pub struct PhysicalMapping {
-    start_frame: PhysFrame<Size4KiB>,
-    start_page: Page<Size4KiB>,
+pub struct PhysicalMapping<S: PageSize = Size4KiB>
+where
+    for<'a> RecursivePageTable<'a>: Mapper<S>,
+{
+    start_frame: PhysFrame<S>,
+    start_page: Page<S>,
     count: u64,
 }
 
-impl PhysicalMapping {
+impl<S: PageSize> PhysicalMapping<S>
+where
+    for<'a> RecursivePageTable<'a>: Mapper<S>,
+{
     /// Creates a new physical mapping.
     ///
     /// `flags` will be `OR`ed with `PageTableFlags::PRESENT` to ensure the page is present.
     #[must_use]
-    pub fn new(start_paddr: PhysAddr, required_length: usize, flags: PageTableFlags) -> Self {
+    pub fn new(start_paddr: PhysAddr, required_length: usize, flags: PageTableFlags) -> Self
+    where
+        S: core::fmt::Debug,
+    {
         let end_paddr = start_paddr + u64::try_from(required_length).unwrap();
 
-        let start_frame = PhysFrame::<Size4KiB>::containing_address(start_paddr);
-        let end_frame = PhysFrame::<Size4KiB>::containing_address(end_paddr);
+        let start_frame = PhysFrame::<S>::containing_address(start_paddr);
+        let end_frame = PhysFrame::<S>::containing_address(end_paddr);
 
         let frame_range = PhysFrame::range_inclusive(start_frame, end_frame);
 
         let count = end_frame - start_frame + 1;
 
         let page_range = page_alloc::with_page_allocator(|page_allocator| {
-            page_allocator.allocate_pages::<Size4KiB>(count).unwrap()
+            page_allocator.allocate_pages::<S>(count).unwrap()
         });
 
         frame_alloc::with_frame_allocator(|frame_allocator| {
@@ -70,7 +81,7 @@ impl PhysicalMapping {
         }
 
         let offset = addr - self.start_frame.start_address();
-        if offset >= self.count * Size4KiB::SIZE {
+        if offset >= self.count * S::SIZE {
             return None;
         }
 
@@ -78,13 +89,16 @@ impl PhysicalMapping {
     }
 }
 
-impl Drop for PhysicalMapping {
+impl<S: PageSize> Drop for PhysicalMapping<S>
+where
+    for<'a> RecursivePageTable<'a>: Mapper<S>,
+{
     fn drop(&mut self) {
         // TODO: Is it possible to add frames to the frame allocator pool at some point?
         // We don't need to keep memory reserved for ACPI once we've read the tables.
         // Be careful as the frame could be used by another mapping.
         let page_range =
-            Page::<Size4KiB>::range_inclusive(self.start_page, self.start_page + self.count - 1);
+            Page::<S>::range_inclusive(self.start_page, self.start_page + self.count - 1);
 
         page_table::with_page_table(|page_table| {
             for page in page_range {
