+++ kernel/src/process/dummy.rs
@@ -4,8 +4,8 @@
 /// The purpose of this thread is to crash after it starts
 /// because of overflow.
 pub fn fibonacci() {
-    let mut a = 0_usize;
-    let mut b = 1_usize;
+    let mut a = 0_u16;
+    let mut b = 1_u16;
     let mut n = 0_u32;
 
     loop {
@@ -20,12 +20,12 @@ pub fn fibonacci() {
         a = b;
         b = next;
         n += 1;
-        crate::time::wait_ms(100);
+        crate::time::wait_ms(1_000);
     }
 }
 
 pub fn counter() {
-    let mut counter = 0_u32;
+    let mut counter = 0_u64;
 
     loop {
         crate::info!(
